% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_methods_expand.R
\name{bout_expand}
\alias{bout_expand}
\alias{bout_expand.default}
\title{Re-vectorize run length encoded data}
\usage{
bout_expand(bouts, ...)

\method{bout_expand}{default}(bouts, ...)
}
\arguments{
\item{bouts}{output from \code{\link{analyze_bouts}}}

\item{...}{arguments passed to methods}
}
\value{
a vector of individual values comprising the runs in \code{bouts}
}
\description{
Re-vectorize run length encoded data
}
\examples{
data(example_data, package = "PBpatterns")
intensity <- cut(
  example_data$PAXINTEN,
  c(-Inf, 101, 760, Inf),
  c("SB", "LPA", "MVPA"),
  right = FALSE
)
\donttest{
bouts <- analyze_bouts(intensity, target = "MVPA", target_buffer = 1)
tail(bout_expand(bouts), 40)
}
}
