% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_bouts__summary_residual_adjust.R
\name{bout_summary_residual_adjust}
\alias{bout_summary_residual_adjust}
\title{Apply residual adjustment to the output of \code{\link{sb_bout_summary}}
and/or \code{\link{mvpa_bout_summary}}}
\usage{
bout_summary_residual_adjust(d, verbose = FALSE)
}
\arguments{
\item{d}{the output of \code{\link{sb_bout_summary}} and/or
\code{\link{mvpa_bout_summary}}}

\item{verbose}{logical. Print updates to console}
}
\value{
The original \code{d}, with extra columns (\code{adj_*}) reflecting
  residual adjustments on relevant variables
}
\description{
Apply residual adjustment to the output of \code{\link{sb_bout_summary}}
and/or \code{\link{mvpa_bout_summary}}
}
\examples{
data(example_data, package = "PBpatterns")

example_data$is_sb <- example_data$PAXINTEN <= 100
example_data$is_mvpa <- example_data$PAXINTEN >= 760
example_data$Timestamp <- seq(
  as.POSIXct("00:00:00", "UTC", format = "\%H:\%M:\%S"),
  by = "1 min",
  length.out = nrow(example_data)
)

valid_indices <- c(
  654:1454, 1917:2837, 3499:4266, 5216:5632,
  6340:7119, 7704:8555, 9118:10077
)

sb_bouts <- sb_bout_summary(example_data, valid_indices = valid_indices)

bout_summary_residual_adjust(sb_bouts)

}
