% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Package_and_Data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Data for use in code examples throughout the \code{PBpatterns} package}
\format{
An object of class \code{data.frame} with 10080 rows and 8 columns.
}
\usage{
example_data
}
\description{
This is a minute-by-minute data set for one participant in the 2003-2006
National Health and Nutrition Examination Survey (NHANES). For details, see the
\href{https://wwwn.cdc.gov/Nchs/Nhanes/2005-2006/PAXRAW_D.htm}{NHANES documentation}.
(Link is to the 2005-2006 documentation, as an additional variable (PAXSTEP)
was included in that cycle.)
}
\keyword{datasets}
