% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_methods_expand.R
\name{expand_bouts}
\alias{expand_bouts}
\alias{expand_bouts.default}
\title{Re-vectorize run length encoded data}
\usage{
expand_bouts(bouts, ...)

\method{expand_bouts}{default}(bouts, ...)
}
\arguments{
\item{bouts}{output from \code{\link{analyze_bouts}}}

\item{...}{arguments passed to methods}
}
\value{
a vector of individual values comprising the runs in \code{bouts}
}
\description{
Re-vectorize run length encoded data
}
\examples{
data(example_data, package = "PBpatterns")
intensity <- cut(
  example_data$PAXINTEN,
  c(-Inf, 101, 760, Inf),
  c("SB", "LPA", "MVPA"),
  right = FALSE
)
bouts <- analyze_bouts(
  intensity, "MVPA", "Troiano_MVPA", epoch_length_sec = 60
)
expand_bouts(bouts)[910:940]
}
