% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibbing21_bouts.R, R/hibbing21_helpers.R
\name{sb_bout_dist_default}
\alias{sb_bout_dist_default}
\alias{sb_bout_dist_df}
\alias{Hibbing_2021_internal}
\alias{df_check_format}
\alias{counts_verify}
\alias{valid_indices_verify}
\alias{id_wear_time_verify}
\alias{id_verify}
\alias{id_bind}
\title{Functions used internally in the package}
\usage{
sb_bout_dist_default(
  is_sb,
  is_wear,
  min_bout = 5,
  valid_indices = NULL,
  probs = c(0.1, 0.2, 0.25, seq(0.3, 0.7, 0.1), 0.75, 0.8, 0.9)
)

sb_bout_dist_df(
  df,
  counts = NULL,
  wear = NULL,
  id = NULL,
  sb = 100,
  min_bout = 5,
  valid_indices = NULL,
  simplify = TRUE,
  probs = c(0.1, 0.2, 0.25, seq(0.3, 0.7, 0.1), 0.75, 0.8, 0.9)
)

df_check_format(df, counts, valid_indices, id, wear)

counts_verify(df, counts)

valid_indices_verify(df, valid_indices)

id_wear_time_verify(df, id, wear)

id_verify(df, id)

id_bind(result, id, simplify = TRUE)
}
\arguments{
\item{is_sb}{logical vector reflecting minute-by-minute classifications
(\code{TRUE} for sedentary behavior and \code{FALSE} for non-sedentary
behavior)}

\item{is_wear}{logical vector reflecting minute-by-minute wear
classifications (\code{TRUE} for wearing and \code{FALSE} for not wearing)}

\item{min_bout}{integer scalar. Threshold for a run to qualify as a bout}

\item{valid_indices}{integer vector (optional) specifying which indices of
\code{is_sb} and {is_wear} correspond to valid measurement days (e.g. with
10+ hours of wear time on 4+ days)}

\item{probs}{numeric vector. Percentiles to calculate (all values must be
between 0 and 1)}

\item{df}{data frame of accelerometer input}

\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior (and wear time, depending on the function)}

\item{wear}{character scalar [optional]. Column name of the variable to use
for determining wear time (logical vector with \code{TRUE} for wear time
minutes). If no value is provided, \code{\link{choi_wear}} is invoked on
the \code{counts} column}

\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{sb}{integer. The cut point to use for classifying sedentary behavior}

\item{simplify}{logical. If passing a data frame, setting to \code{FALSE}
will return separate bout objects that can be passed into
\code{\link{sb_profile}}). Setting to \code{TRUE} will return an
aggregated data frame through \code{rbind} and \code{\link{id_bind}}.}

\item{result}{output that may need data frame formatting (possibly ID-based)}
}
\description{
Functions used internally in the package
}
\details{
The purpose of \code{id_wear_time_verify} is to determine if the Choi
  wear time algorithm needs to be run and, if so, to run it (via
  \code{\link{choi_wear}}, with a message). To bypass,
  run \code{df$is_wear <- TRUE} prior to executing
  \code{id_wear_time_verify(df, id, wear = "is_wear")}.

  \code{df_check_format} is a wrapper that encompasses
}
\keyword{internal}
