% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_PBpatterns.R
\docType{package}
\name{PBpatterns}
\alias{PBpatterns}
\title{Analyze Patterns of Physical Behavior}
\description{
Useful for assessing sedentary patterns and profiles, bouts of
  physical activity or sleep, and more.
}
\seealso{
https://github.com/paulhibbing/PBpatterns
}
