% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bout_methods_plot.R
\name{plot.bouts}
\alias{plot.bouts}
\title{Visualize the result of a bout analysis}
\usage{
\method{plot}{bouts}(x, ...)
}
\arguments{
\item{x}{the object to plot}

\item{...}{unused}
}
\value{
A \code{ggplot2} object that visualizes the bout analysis
}
\description{
Visualize the result of a bout analysis
}
\examples{
data(example_data, package = "PBpatterns")
intensity <- cut(
  example_data$PAXINTEN,
  c(-Inf, 101, 760, Inf),
  c("SB", "LPA", "MVPA"),
  right = FALSE
)
bouts <- analyze_bouts(
  intensity, "MVPA", "Troiano_MVPA", epoch_length_sec = 60
)
# plot(bouts)
}
