% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_profile__Hibbing2021d__SB.R
\name{profile_describe_sb}
\alias{profile_describe_sb}
\alias{profile_describe_sb_df}
\title{Extract a summary of bout duration distribution for sedentary profiles}
\usage{
profile_describe_sb(
  df = NULL,
  is_sb,
  is_wear,
  minimum_bout_length = 5,
  valid_indices = NULL,
  id = NULL,
  counts = NULL,
  wear = NULL,
  sb = 100,
  simplify = TRUE,
  probs = c(0.1, 0.2, 0.25, seq(0.3, 0.7, 0.1), 0.75, 0.8, 0.9),
  epoch_length_sec
)

profile_describe_sb_df(
  df,
  counts = NULL,
  wear = NULL,
  id = NULL,
  sb = 100,
  minimum_bout_length = 5,
  valid_indices = NULL,
  simplify = TRUE,
  probs = c(0.1, 0.2, 0.25, seq(0.3, 0.7, 0.1), 0.75, 0.8, 0.9),
  epoch_length_sec
)
}
\arguments{
\item{df}{optional data frame input (possibly with data for multiple strata,
e.g., different participants)}

\item{is_sb}{logical vector indicating epoch-by-epoch classifications of
sedentary behavior (\code{TRUE}) or non-sedentary (\code{FALSE})}

\item{is_wear}{logical vector indicating epoch-by-epoch classifications of
wear time (\code{TRUE}) or non-wear  (\code{FALSE})}

\item{minimum_bout_length}{numeric. Minimum number of epochs in a valid bout}

\item{valid_indices}{[optional] numeric/integer/logical vector specifying
which indices of \code{is_sb} and {is_wear} correspond to valid measurement
days (e.g. with 10+ hours of wear time on 4+ days). The default
(\code{NULL}) assumes all elements are valid}

\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior (and wear time, depending on the function)}

\item{wear}{character scalar [optional]. Column name of the variable to use
for determining wear time (logical vector with \code{TRUE} for wear time
minutes). If no value is provided, \code{\link{profile_choi_wear}} is invoked on
the \code{counts} column}

\item{sb}{numeric. The cut point to use for classifying sedentary behavior}

\item{simplify}{logical. Collapse output for each stratum into a single data
frame?}

\item{probs}{numeric vector. Quantiles to summarize in the output}

\item{epoch_length_sec}{numeric. The epoch_length of \code{x}}
}
\description{
Extract a summary of bout duration distribution for sedentary profiles
}
\details{
\code{profile_describe_sb} will "dispatch" to
  \code{profile_describe_sb_df} if a value is provided for \code{df}.
  Arguments can be used accordingly, but should always be named.
}
\examples{
data(example_data, package = "PBpatterns")
example_data$is_wear <- TRUE
profile_describe_sb(
  example_data, counts = "PAXINTEN",
  wear = "is_wear", id = "PAXDAY",
  epoch_length_sec = 60
)
}
\keyword{internal}
