% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularity__generic.R, R/regularity_factor.R,
%   R/regularity_numeric.R
\name{regularity_index}
\alias{regularity_index}
\alias{regularity_index.factor}
\alias{regularity_index.numeric}
\title{Calculate regularity indices for analyzing time-of-day consistency of
recurring behaviors}
\usage{
regularity_index(values, timestamps, ...)

\method{regularity_index}{factor}(values, timestamps, target, min_days = 3, numeric = TRUE, ...)

\method{regularity_index}{numeric}(
  values,
  timestamps,
  n_bands,
  band_size,
  min_days = 3,
  numeric = TRUE,
  ...
)
}
\arguments{
\item{values}{a vector of values to analyze}

\item{timestamps}{a vector of \code{POSIX-formatted} timestamps. Must have
the same length as \code{values}, and it must have element-wise, one-to-one
correspondence with the elements of \code{values}}

\item{...}{further arguments passed to methods}

\item{target}{character scalar. When \code{values} is a factor, which level
should be emphasized? All other levels will be collapsed into a single
'other' category.}

\item{min_days}{the minimum number of days that are required to ensure proper
analysis for each time of day. A warning will issue for any daytimes (i.e.,
epochs) that fall below this threshold.}

\item{numeric}{logical. Return the score as a numeric object (default)? If
\code{FALSE}, a \code{regularity} object will be returned instead.}

\item{n_bands}{the number of difference bands that will be examined for a
continuous input}

\item{band_size}{the size of each difference band that will be examined for a
continuous input}
}
\value{
A regularity index whose value lies between -100 and +100.
}
\description{
Calculate regularity indices for analyzing time-of-day consistency of
recurring behaviors
}
\details{
When setting \code{numeric = TRUE}, only the numeric value is
returned. Otherwise, a \code{regularity} object is returned, whose value
equals the numeric value, and whose attributes include additional information
from the analysis (potentially useful for debugging).
}
\examples{
set.seed(1919)
values <- factor(
  sample(c("Awake", "Asleep"), 6, TRUE)
)
timestamp_offsets <- 43200 * 0:5
regularity_index(values, as.POSIXlt(Sys.Date()) + timestamp_offsets, min_days = 2)
regularity_index(values, as.POSIXlt(Sys.Date()) + timestamp_offsets, min_days = 2, numeric = FALSE)
}
