% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle_bouts__internal.R
\name{rle_bouts_internal}
\alias{rle_bouts_internal}
\alias{validate_rle_bout_input}
\alias{rle_bouts}
\alias{validate_rle_bouts}
\alias{other_info_weartime}
\title{Internal functions for summarizing bouts}
\usage{
validate_rle_bout_input(d, is_activity, is_wear, time_var)

rle_bouts(
  d,
  is_activity,
  is_wear,
  epoch_length_sec,
  min_bout_minutes = 0,
  valid_indices = 1:nrow(d)
)

validate_rle_bouts(bouts, valid_indices)

other_info_weartime(
  other_info,
  d,
  is_wear,
  time_var,
  valid_indices,
  epoch_length_sec
)
}
\description{
Internal functions for summarizing bouts
}
\keyword{internal}
