% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibbing21_bouts.R
\name{sb_bout_dist}
\alias{sb_bout_dist}
\title{Extract features to summarize a participant's bout duration distribution}
\usage{
sb_bout_dist(
  df = NULL,
  is_sb,
  is_wear,
  min_bout = 5,
  valid_indices = NULL,
  id = NULL,
  counts = NULL,
  wear = NULL,
  sb = 100,
  simplify = TRUE,
  probs = c(0.1, 0.2, 0.25, seq(0.3, 0.7, 0.1), 0.75, 0.8, 0.9)
)
}
\arguments{
\item{df}{data frame of accelerometer input}

\item{is_sb}{logical vector reflecting minute-by-minute classifications
(\code{TRUE} for sedentary behavior and \code{FALSE} for non-sedentary
behavior)}

\item{is_wear}{logical vector reflecting minute-by-minute wear
classifications (\code{TRUE} for wearing and \code{FALSE} for not wearing)}

\item{min_bout}{integer scalar. Threshold for a run to qualify as a bout}

\item{valid_indices}{integer vector (optional) specifying which indices of
\code{is_sb} and {is_wear} correspond to valid measurement days (e.g. with
10+ hours of wear time on 4+ days)}

\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior (and wear time, depending on the function)}

\item{wear}{character scalar [optional]. Column name of the variable to use
for determining wear time (logical vector with \code{TRUE} for wear time
minutes). If no value is provided, \code{\link{choi_wear}} is invoked on
the \code{counts} column}

\item{sb}{integer. The cut point to use for classifying sedentary behavior}

\item{simplify}{logical. If passing a data frame, setting to \code{FALSE}
will return separate bout objects that can be passed into
\code{\link{sb_profile}}). Setting to \code{TRUE} will return an
aggregated data frame through \code{rbind} and \code{\link{id_bind}}.}

\item{probs}{numeric vector. Percentiles to calculate (all values must be
between 0 and 1)}
}
\value{
A bout object (data frame of distribution features)
}
\description{
Extract features to summarize a participant's bout duration distribution
}
\details{
This function can be used in one of two ways, either 1) by directly
  providing values for \code{is_sb} and \code{is_wear} (with \code{df =
  NULL}), or 2) by providing values for \code{df} and \code{counts}
  (optionally with \code{wear} as well).
}
\examples{
data(example_data, package = "PBpatterns")
sb_bout_dist(
  is_sb = example_data$PAXINTEN <= 100,
  is_wear = choi_wear(example_data$PAXINTEN)
)
sb_bout_dist(
  example_data, id = "PAXDAY", counts = "PAXINTEN"
)
}
