% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_pattern_functions.R
\name{SB_bouts_internal}
\alias{SB_bouts_internal}
\alias{cut_bouts}
\alias{usual_bout_duration}
\alias{fragmentation_index}
\alias{gini}
\alias{alpha}
\title{Internal functions providing information about sedentary bout patterns}
\usage{
cut_bouts(d, bouts)

usual_bout_duration(d, bouts)

fragmentation_index(d, total_length, bouts)

gini(d, bouts)

alpha(d, bouts)
}
\arguments{
\item{d}{one-row data frame with bout information}

\item{bouts}{data frame with information for each bout, based on
\code{PAutilities::index_runs}}

\item{total_length}{the number of rows in the original data.}
}
\description{
These functions are invoked in the \code{\link{sb_bout_summary}} wrapper
}
\note{
There is some potential for confusion, because \code{a} is equivalent
  to the \code{d} argument of \code{\link{sb_bout_summary}}, while \code{d}
  is a one-row data frame generated by the initial operations of
  \code{\link{sb_bout_summary}}. This would be avoidable with revision, but
  there is not much incentive for doing so because this is an internal
  function.
}
\keyword{internal}
