% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_patterns__main.R, R/sb_patterns_doc.R
\name{SB_patterns}
\alias{SB_patterns}
\alias{sb_patterns}
\alias{sb_range_bouts}
\alias{usual_bout_duration}
\alias{fragmentation_index}
\alias{gini}
\alias{alpha}
\title{Internal functions providing information about sedentary bout patterns}
\usage{
sb_patterns(d, bouts, x, epoch_length_sec)

sb_range_bouts(d, bouts, epoch_length_sec)

usual_bout_duration(d, bouts)

fragmentation_index(d, bouts, x)

gini(d, bouts)

alpha(d, bouts)
}
\arguments{
\item{d}{one-row data frame with bout information}

\item{bouts}{data frame with information for each bout, based on
\code{PAutilities::index_runs}}

\item{x}{original data stream}

\item{epoch_length_sec}{numeric. The epoch length of \code{x}}
}
\description{
These functions are invoked in the \code{\link{analyze_bouts}} when
\code{method == "SB_summary"} unless suppressed with an additional
\code{patterns = FALSE} argument.
}
\keyword{internal}
