% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_profile.R
\name{sb_profile}
\alias{sb_profile}
\title{Implement the sedentary profiles}
\usage{
sb_profile(object, method = "Hibbing_2021", ...)
}
\arguments{
\item{object}{input (either data frame or output from
\code{\link{sb_bout_dist}})}

\item{method}{character. The profile method to use (currently only supports
\code{Hibbing_2021}).}

\item{...}{further arguments passed to \code{\link{sb_bout_dist}} (used only
in data frame method)}
}
\value{
one or more profile assignments
}
\description{
Implement the sedentary profiles
}
\examples{
data(example_data, package = "PBpatterns")
sb_profile(example_data, counts = "PAXINTEN")
sb_profile(
  example_data, counts = "PAXINTEN", model = "randomForest"
)
sb_profile(example_data, id = "PAXDAY", counts = "PAXINTEN")
sb_profile(
  example_data, id = "PAXDAY",
  counts = "PAXINTEN", model = "decisionTree"
)

}
\seealso{
\code{\link{sb_profile_Hibbing2021}}
  \href{https://journals.lww.com/acsm-msse/Abstract/9000/Sedentary_Profiles__A_New_Perspective_on.95921.aspx}{Hibbing et al. (2021)}
}
