% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_profile.R
\name{sb_profile_Hibbing2021}
\alias{sb_profile_Hibbing2021}
\alias{sb_profile_Hibbing2021.bout}
\alias{sb_profile_Hibbing2021.data.frame}
\alias{sb_profile_Hibbing2021.list}
\title{Apply the SB profiles of Hibbing et al. (2021)}
\usage{
sb_profile_Hibbing2021(object, method = "Hibbing_2021", ...)

\method{sb_profile_Hibbing2021}{bout}(
  object,
  method = "Hibbing_2021",
  model = c("both", "decisionTree", "randomForest"),
  ...
)

\method{sb_profile_Hibbing2021}{data.frame}(
  object,
  method = "Hibbing_2021",
  model = c("both", "decisionTree", "randomForest"),
  id = NULL,
  counts = NULL,
  wear = NULL,
  sb = 100,
  valid_indices = NULL,
  ...
)

\method{sb_profile_Hibbing2021}{list}(
  object,
  method = "Hibbing_2021",
  model = c("both", "decisionTree", "randomForest"),
  ...
)
}
\arguments{
\item{object}{input (either data frame or output from
\code{\link{sb_bout_dist}})}

\item{method}{character. The profile method to use (currently only supports
\code{Hibbing_2021}).}

\item{...}{further arguments passed to \code{\link{sb_bout_dist}} (used only
in data frame method)}

\item{model}{character. The model(s) to use for the prediction}

\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior (and wear time, depending on the function)}

\item{wear}{character scalar [optional]. Column name of the variable to use
for determining wear time (logical vector with \code{TRUE} for wear time
minutes). If no value is provided, \code{\link{choi_wear}} is invoked on
the \code{counts} column}

\item{sb}{integer. The cut point to use for classifying sedentary behavior}

\item{valid_indices}{integer vector (optional) specifying which indices of
\code{is_sb} and {is_wear} correspond to valid measurement days (e.g. with
10+ hours of wear time on 4+ days)}
}
\description{
Apply the SB profiles of Hibbing et al. (2021)
}
