% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_weartime.R
\name{summarize_weartime}
\alias{summarize_weartime}
\title{Generate a summary of accelerometer wear time}
\usage{
summarize_weartime(
  d,
  is_wear,
  time_var,
  valid_indices = NULL,
  other_info = NULL
)
}
\arguments{
\item{d}{data frame of accelerometer data}

\item{is_wear}{character. Column name of a logical variable indicating
whether each epoch is wear time (\code{TRUE}) or not (\code{FALSE})}

\item{time_var}{character. Column name of a timestamp variable}

\item{valid_indices}{optional vector (logical or integer/numeric) indicating
which rows in \code{d} correspond with valid wear days. If no value is
provided, all rows will be considered valid}

\item{other_info}{optional one-row data frame containing additional
information to bind with the results of wear time summarizing}
}
\value{
A one-row data frame containing the wear time summary and anything
  else passed in through \code{other_info}
}
\description{
Generate a summary of accelerometer wear time
}
\examples{
set.seed(610)
data(example_data, package = "PBpatterns")

example_data$timestamp <- seq(
  Sys.time(), by = "1 min", length.out = nrow(example_data)
)
example_data$is_wear <- sample(c(FALSE, TRUE), nrow(example_data), TRUE)
example_data$intensity <- cut(
  example_data$PAXINTEN,
  c(-Inf, 101, 760, Inf),
  c("SB", "LPA", "MVPA"),
  right = FALSE
)

extra_info <- analyze_bouts(
  example_data$intensity, "SB", "SB_summary",
  is_wear = example_data$is_wear, epoch_length_sec = 60
)

## Standalone usage
summarize_weartime(example_data, "is_wear", "timestamp")

## In conjunction with extra information
summarize_weartime(
  example_data, "is_wear", "timestamp",
  other_info = extra_info
)
}
