% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_bouts.R, R/troiano_mvpa_bouts.R
\name{analyze_bouts}
\alias{analyze_bouts}
\alias{troiano_mvpa_bouts}
\title{Analyze bouts of various physical behaviors}
\usage{
analyze_bouts(x, target, method = c("CRIB", "Troiano_MVPA"), ...)

troiano_mvpa_bouts(
  x,
  target,
  activation_window = 10,
  activation_min = 8,
  termination_min = 3
)
}
\arguments{
\item{x}{factor. Epoch-by-epoch behavior classifications, e.g., \code{SB},
\code{LPA}, and \code{MVPA}}

\item{target}{the level of \code{x} for which to examine bouts}

\item{method}{character. The bout classifying method to use. See details.}

\item{...}{arguments passed to the method functions being wrapped}

\item{activation_window}{size of window (number of epochs) to use when
searching for a bout activation}

\item{activation_min}{number of epochs in the activation window that must
equal \code{target} for an activation to be detected}

\item{termination_min}{number of consecutive non-\code{target} epochs
required to terminate a bout}
}
\description{
This is a universal function that serves as a wrapper for several bout
classifying methods.
}
\details{
Currently, two methods are supported, namely \code{"CRIB"} and
  \code{"Troiano_MVPA"}. More can easily be added over time, including
  those designed for specific behaviors or activity monitors.
}
\examples{
data(ex_data, package = "PAutilities")
}
\seealso{
\href{https://journals.lww.com/acsm-msse/pages/articleviewer.aspx?year=2008&issue=01000&article=00025&type=Fulltext}{Troiano et al. (2008)}
}
\keyword{internal}
