% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_profile__Hibbing2021b__internal_doc.R,
%   R/sb_profile__Hibbing2021c__profile_checks.R,
%   R/sb_profile__Hibbing2021e__ID_bind.R
\name{Hibbing_2021_internal}
\alias{Hibbing_2021_internal}
\alias{profile_df_check}
\alias{profile_counts_check}
\alias{profile_indices_check}
\alias{profile_id_wear_check}
\alias{profile_id_verify}
\alias{profile_choi_wear}
\alias{profile_id_bind}
\title{Functions used internally for the sedentary profiles of \href{https://journals.lww.com/acsm-msse/Abstract/9000/Sedentary_Profiles__A_New_Perspective_on.95921.aspx}{Hibbing et al. (2021)}}
\usage{
profile_df_check(df, counts, valid_indices, id, wear)

profile_counts_check(df, counts)

profile_indices_check(df, valid_indices)

profile_id_wear_check(df, id, wear)

profile_id_verify(df, id)

profile_choi_wear(counts)

profile_id_bind(result, id, simplify = TRUE)
}
\arguments{
\item{counts}{character scalar. Column name of the variable to use when
classifying sedentary behavior (and wear time, depending on the function)}

\item{valid_indices}{integer vector (optional) specifying which indices of
\code{is_sb} and {is_wear} correspond to valid measurement days (e.g. with
10+ hours of wear time on 4+ days)}

\item{id}{character scalar (optional). Column name on which to divide
\code{object} (if a data frame) into a list of separate objects}

\item{wear}{character scalar [optional]. Column name of the variable to use
for determining wear time (logical vector with \code{TRUE} for wear time
minutes). If no value is provided, \code{\link{profile_choi_wear}} is invoked on
the \code{counts} column}

\item{result}{output that may need data frame formatting (possibly ID-based)}
}
\description{
Functions used internally for the sedentary profiles of \href{https://journals.lww.com/acsm-msse/Abstract/9000/Sedentary_Profiles__A_New_Perspective_on.95921.aspx}{Hibbing et al. (2021)}
}
\details{
The purpose of \code{profile_id_wear_check} is to determine if the Choi
  wear time algorithm needs to be run and, if so, to run it (via
  \code{profile_choi_wear}, with a message). To bypass,
  run \code{df$is_wear <- TRUE} prior to executing
  \code{profile_id_wear_check(df, id, wear = "is_wear")}.

  \code{profile_df_check} is a wrapper that encompasses all the checks on this
  page
}
\keyword{internal}
